refactor: Presenter 분리
