+++ app/src/main/java/woowacourse/movie/view/reservation/seat/SeatSelectionContract.kt
@@ -2,8 +2,6 @@ package woowacourse.movie.view.reservation.seat
 
 import woowacourse.movie.domain.model.ReservationInfo
 import woowacourse.movie.domain.model.Seat
-import woowacourse.movie.domain.model.SeatFactory
-import woowacourse.movie.domain.model.TicketMachine
 
 interface SeatSelectionContract {
     interface View {
@@ -35,54 +33,5 @@ interface SeatSelectionContract {
         fun showConfirmButton()
 
         fun completeReservation()
-    }
-}
-
-class SeatSelectionPresenter(
-    private val view: SeatSelectionContract.View,
-) : SeatSelectionContract.Presenter {
-    private var reservationInfo: ReservationInfo? = null
-
-    private val seatFactory = SeatFactory.default()
-    private val ticketMachine = TicketMachine()
-
-    override fun loadSeats(reservationInfo: ReservationInfo?) {
-        val seats = seatFactory.createSeats()
-        this.reservationInfo = reservationInfo
-
-        view.showSeats(seats)
-        updateScreen()
-    }
-
-    override fun selectSeat(seat: Seat) {
-        try {
-            reservationInfo?.updateSeats(seat)
-            view.updateSeatSelection(seat, seat.isSelected)
-            updateScreen()
-        } catch (e: IllegalArgumentException) {
-            view.showError(e.message ?: "좌석 선택 오류")
-        }
-    }
-
-    private fun updateScreen() {
-        view.showTotalPrice(reservationInfo?.totalPrice() ?: throw IllegalArgumentException())
-        view.enableConfirmButton(canCompleteReservation())
-    }
-
-    private fun canCompleteReservation(): Boolean = reservationInfo?.let { ticketMachine.canPublish(it) } ?: false
-
-    override fun showConfirmButton() {
-        if (canCompleteReservation()) {
-            view.showReservationDialog()
-        } else {
-            view.showError("좌석을 선택해주세요")
-        }
-    }
-
-    override fun completeReservation() {
-        reservationInfo?.let {
-            ticketMachine.publishTickets(it)
-            view.navigateToResult(it)
-        }
     }
 }
